refactor(#3/user): user 도메인 내부 엔티티, enum 패키지 분리
