+++ src/main/java/org/sopt/snappinserver/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package org.sopt.snappinserver.domain.user.repository;
 
-import org.sopt.snappinserver.domain.user.domain.User;
+import org.sopt.snappinserver.domain.user.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
